requirements.txt
# FastAPI backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# Database
aiosqlite==0.20.0

# AI/ML
openai==1.54.3
httpx==0.27.2

# Frontend
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.3

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0

# Data scraping
beautifulsoup4==4.12.3
lxml==5.3.0

